requirements.txt
Flask==3.0.1
flask_cors==4.0.0
duckduckgo_search==4.2
newspaper3k
simplejson
